KTB-15-Send-a-message-to-a-user-via-Telegram-API refactoring total
