+++ src/main/kotlin/Telegram.kt
@@ -1,63 +1,41 @@
 package org.example
 
-import java.net.URI
-import java.net.URLEncoder
-import java.net.http.HttpClient
-import java.net.http.HttpRequest
-import java.net.http.HttpResponse
-
-class TelegramBotService(val botToken: String) {
-    val client: HttpClient = HttpClient.newBuilder().build()
-
-    fun getUpdates(updateId: Int): String {
-        val urlGetUpdates = "https://api.telegram.org/bot$botToken/getUpdates?offset=${this}"
-        val request: HttpRequest? = HttpRequest.newBuilder().uri(URI.create(urlGetUpdates)).build()
-        val response: HttpResponse<String?> = client.send(request, HttpResponse.BodyHandlers.ofString())
-
-        return response.body().toString()
-    }
-
-    fun String.encodeUrl(): String? = URLEncoder.encode(this, Charsets.UTF_8.name())
-
-    fun sendMessage(chatId: Long, text: String): String? {
-        val encoded = text.encodeUrl()
-        val urlSendMess = "https://api.telegram.org/bot${this}/sendMessage?chat_id=$chatId&text=$encoded"
-        val request: HttpRequest? = HttpRequest.newBuilder().uri(URI.create(urlSendMess)).build()
-        val response: HttpResponse<String?> = client.send(request, HttpResponse.BodyHandlers.ofString())
-
-        return response.body()
-    }
-}
-
 fun main(args: Array<String>) {
 
     val botToken = args[0]
     var updateId = 0
     val telegramBotService = TelegramBotService(botToken)
+    val updateIdRegex: Regex = "\"update_id\":([0-9]+)[,}]".toRegex()
+    val messageTextRegex: Regex = "\"text\":\"(.+?)\"".toRegex()
+    val chatIdRegex: Regex = "\"chat\":\\{\"id\":(\\d+)".toRegex()
 
     while (true) {
         Thread.sleep(2000)
         val updates: String = telegramBotService.getUpdates(updateId)
         println(updates)
 
-        val startUpdateId = updates.lastIndexOf("update_id")
-        val endUpdateId = updates.lastIndexOf(",\n\"message\"")
-        if (startUpdateId == -1 || endUpdateId == -1) continue
-        val updateIdString = updates.substring(startUpdateId + 11, endUpdateId)
-        println(updateIdString)
-        updateId = updateIdString.toInt() + 1
+        val updateIdGroups = updateIdRegex.find(updates)?.groups
+        val newUpdateId = updateIdGroups?.get(1)?.value?.toInt() ?: 0
+        updateId = newUpdateId + 1
 
-        val messageTextRegex: Regex = "\"text\":\"(.+?)\"".toRegex()
+
         val matchResult: MatchResult? = messageTextRegex.find(updates)
         val groups = matchResult?.groups
-        val text = groups?.get(1)?.value
+        val text = groups?.get(1)?.value?.decodeUnicode()
         println(text)
 
-        val chatIdRegex: Regex = "\"chat\":\\{\"id\":(\\d+)".toRegex()
         val matchResult1: MatchResult? = chatIdRegex.find(updates)
         val groups1: MatchGroupCollection? = matchResult1?.groups
         val chatId: Long = groups1?.get(1)?.value?.toLongOrNull() ?: continue
 
         telegramBotService.sendMessage(chatId, "$text")
+    }
+}
+
+fun String.decodeUnicode(): String {
+    return this.replace("""\\u([0-9a-fA-F]{4})""".toRegex()) { matchResult ->
+        val hexCode = matchResult.groupValues[1]
+        val codePoint = hexCode.toInt(16)
+        String(Character.toChars(codePoint))
     }
 }
